// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  admin: typeof admin;
  adminActions: typeof adminActions;
  auth: typeof auth;
  email: typeof email;
  emailActions: typeof emailActions;
  emailHelpers: typeof emailHelpers;
  http: typeof http;
  stripe: typeof stripe;
  users: typeof users;
}>;
// ...
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;